+++ package/Tests/ClassA.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+
+namespace Singletons.Tests
+{
+	public class ClassA : Singleton<ClassA>
+	{
+		
+	}
+}