+++ package/Tests/ClassC.cs
@@ -1,0 +1,7 @@
+namespace Singletons.Tests
+{
+	public class ClassC : ClassA
+	{
+
+	}
+}